Move key def sections into a subdirectory to make the codebase a bit cleaner.
These should never have been getting called directly by end-user code, although the URLs were referenced as a resource in docs.
